+++ src/sevDesk.NET/Internal/ModelMapper.cs
@@ -62,6 +62,7 @@ internal static class ModelMapper
         Birthday = ParseDateTime(api.Birthday),
         DefaultDiscountPercentage = ParseDecimal(api.DefaultDiscountPercentage),
         GovernmentAgency = ParseBool(api.GovernmentAgency),
+        BuyerReference = api.BuyerReference,
         Create = ParseDateTime(api.Create),
         Update = ParseDateTime(api.Update)
     };
@@ -91,7 +92,8 @@ internal static class ModelMapper
         ExemptVat = FormatBool(model.ExemptVat),
         Birthday = FormatDateTime(model.Birthday),
         DefaultDiscountPercentage = model.DefaultDiscountPercentage?.ToString(System.Globalization.CultureInfo.InvariantCulture),
-        GovernmentAgency = FormatBool(model.GovernmentAgency)
+        GovernmentAgency = FormatBool(model.GovernmentAgency),
+        BuyerReference = model.BuyerReference
     };
 
     // --- Invoice ---
@@ -114,14 +116,14 @@ internal static class ModelMapper
         ContactPerson = ToPublic(api.ContactPerson),
         Address = api.Address,
         Currency = api.Currency,
-        SumNet = ParseDecimal(api.SumNet),
-        SumGross = ParseDecimal(api.SumGross),
-        SumTax = ParseDecimal(api.SumTax),
+        SumNet = api.SumNet,
+        SumGross = api.SumGross,
+        SumTax = api.SumTax,
         TaxType = api.TaxType,
         TaxRate = api.TaxRate,
         TaxText = api.TaxText,
         SendDate = ParseDateTime(api.SendDate),
-        PaymentMethod = api.PaymentMethod,
+        PaymentMethod = ToPublic(api.PaymentMethod),
         CostCentre = ToPublic(api.CostCentre),
         SendType = api.SendType,
         Origin = ToPublic(api.Origin),
@@ -129,7 +131,10 @@ internal static class ModelMapper
         SmallSettlement = ParseBool(api.SmallSettlement),
         TaxSet = ToPublic(api.TaxSet),
         Create = ParseDateTime(api.Create),
-        Update = ParseDateTime(api.Update)
+        Update = ParseDateTime(api.Update),
+        TaxRule = ToPublic(api.TaxRule),
+        EinvoiceReference = api.EinvoiceReference,
+        PropertyIsEInvoice = ParseBool(api.PropertyIsEInvoice)
     };
 
     internal static ApiInvoice ToApi(Invoice model) => new()
@@ -154,13 +159,16 @@ internal static class ModelMapper
         TaxRate = model.TaxRate,
         TaxText = model.TaxText,
         SendDate = FormatDateTime(model.SendDate),
-        PaymentMethod = model.PaymentMethod,
+        PaymentMethod = ToApi(model.PaymentMethod),
         CostCentre = ToApi(model.CostCentre),
         SendType = model.SendType,
         Origin = ToApi(model.Origin),
         CustomerInternalNote = model.CustomerInternalNote,
         SmallSettlement = FormatBool(model.SmallSettlement),
-        TaxSet = ToApi(model.TaxSet)
+        TaxSet = ToApi(model.TaxSet),
+        TaxRule = ToApi(model.TaxRule),
+        EinvoiceReference = model.EinvoiceReference,
+        PropertyIsEInvoice = FormatBool(model.PropertyIsEInvoice)
     };
 
     // --- InvoicePos ---